requirements.txt
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
cmake==3.25.0
comm==0.2.2
ConfigArgParse==1.7.1
contourpy==1.3.0
cycler==0.12.1
Cython==3.1.2
dash==3.1.0
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
detectron2==0.6
easydict==1.13
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
fairscale==0.4.13
fastjsonschema==2.21.1
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
fvcore==0.1.5.post20221221
gdown==5.2.0
gin-config==0.5.0
gitdb==4.0.12
GitPython==3.1.44
h5py==3.14.0
hf-xet==1.1.5
huggingface-hub==0.33.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.14.3
lit==15.0.7
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mmdet==2.26.0
mpmath==1.3.0
multidict==6.6.2
narwhals==1.44.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.4
nltk==3.9.1
numpy==1.24.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
open-clip-torch==2.12.0
open3d==0.18.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openi==2.0.4
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.3.8
plotly==6.2.0
pluggy==1.6.0
portalocker==3.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pwinput==1.0.3
pycocotools==2.0.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pyquaternion==0.9.9
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
retrying==1.4.0
rpds-py==0.25.1
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
sentencepiece==0.2.0
sentry-sdk==2.32.0
setproctitle==1.3.6
shapely==2.0.7
simsimd==6.4.9
six==1.17.0
smmap==5.0.2
soupsieve==2.7
stack-data==0.6.3
stringzilla==3.12.5
sympy==1.13.1
tabulate==0.9.0
termcolor==3.1.0
terminaltables==3.1.10
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==1.0.16
tokenizers==0.13.3
tomli==2.2.1
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchmetrics==1.7.3
torchvision==0.15.1+cu118
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.30.2
triton==2.0.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.20.1
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
yacs==0.1.8
yapf==0.43.0
yarl==1.20.1
zipp==3.23.0
